requirements.txt
paddlepaddle==3.2.1
paddleocr[doc-parser]
streamlit
pandas
Pillow